+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/client/GameRendererMixin.java
@@ -1,12 +1,17 @@
 package insane96mcp.iguanatweaksreborn.mixin.client;
 
 import insane96mcp.iguanatweaksreborn.module.client.Light;
+import insane96mcp.iguanatweaksreborn.module.client.Misc;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.damagesource.DamageTypes;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.LivingEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(GameRenderer.class)
@@ -21,5 +26,18 @@ public abstract class GameRendererMixin {
 		callback.setReturnValue(duration > Light.NIGHT_VISION_FADE_OUT_AT || duration == -1 ?
 				1.0f :
 				((float)duration - partialTicks) * (1f / Light.NIGHT_VISION_FADE_OUT_AT));
+	}
+
+	@ModifyVariable(method = "bobHurt", at = @At(value = "STORE"), ordinal = 3)
+	public float onTiltStrength(float tilt) {
+		if (!Misc.shouldDisableTiltingWithSomeDamageTypes())
+			return tilt;
+		LivingEntity livingEntity = (LivingEntity) Minecraft.getInstance().cameraEntity;
+		DamageSource lastDamageSource = livingEntity.getLastDamageSource();
+		if (lastDamageSource == null)
+			return tilt;
+		if (lastDamageSource.is(DamageTypes.MAGIC) || lastDamageSource.is(DamageTypes.WITHER) || lastDamageSource.is(DamageTypes.ON_FIRE) || lastDamageSource.is(DamageTypes.CRAMMING) || lastDamageSource.is(DamageTypes.THORNS) || lastDamageSource.is(DamageTypes.DROWN))
+			return 0f;
+		return tilt;
 	}
 }