+++ src/main/java/insane96mcp/iguanatweaksreborn/module/hungerhealth/nohunger/NoHunger.java
@@ -26,6 +26,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.util.Mth;
 import net.minecraft.world.Difficulty;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
@@ -393,8 +394,9 @@ public class NoHunger extends Feature {
             float regenStrength = getFoodRegenStrength(player) * 20 * 2;
             if (regenStrength == 0f)
                 return;
+            int width = Mth.ceil(regenLeft / 2f * 9);
             ClientUtils.setRenderColor(1.2f - (regenStrength / 1.2f), 0.78f, 0.17f, 1f);
-            guiGraphics.blit(OT_REGEN_LOCATION, right, top, 0f, 0f, (int) (regenLeft / 2f * 9), 3, 90, 3);
+            guiGraphics.blit(OT_REGEN_LOCATION, right, top, 0f, 0f, width, 3, 90, 3);
             ClientUtils.resetRenderColor();
         });
     }